pyproject.toml
[tool.poetry]
name = "python-postgres-template"
version = "0.1.0"
description = ""
authors = ["template-author"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.7"
python-dotenv = "^1.0.0"
flask = "^2.3.2"
flask-restful = "^0.3.10"
marshmallow = "^3.20.1"
requests = "^2.31.0"
python-dateutil = "^2.8.2"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
coverage = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
